+++ src/baobab_activity_reporting/reporting/__init__.py
@@ -11,8 +11,15 @@ from baobab_activity_reporting.reporting.section_eligibility_evaluator import (
     SectionEligibilityEvaluator,
 )
 from baobab_activity_reporting.reporting.table_builder import TableBuilder
+from baobab_activity_reporting.reporting.writers.abstract_writer import AbstractWriter
+from baobab_activity_reporting.reporting.writers.docx_writer import DocxWriter
+from baobab_activity_reporting.reporting.writers.markdown_writer import (
+    MarkdownWriter,
+)
 
 __all__: list[str] = [
+    "AbstractWriter",
+    "DocxWriter",
     "InsightBuilder",
     "NarrativeBuilder",
     "ReportBuilder",
@@ -21,5 +28,6 @@ __all__: list[str] = [
     "ReportModel",
     "ReportPlanner",
     "SectionEligibilityEvaluator",
+    "MarkdownWriter",
     "TableBuilder",
 ]